+++ src/Blog/elements/PostAbstract/PostAbstract.component.tsx
@@ -1,0 +1,7 @@
+import * as markdown from '../../../Shared/lib/markdown';
+
+import { MarkdownComponentProps, markdownFactory } from '../../../Shared/elements/Markdown/Markdown.factory';
+
+const createComponent = () => markdownFactory(markdown.strippedLinks(), {});
+
+export const PostAbstract: React.StatelessComponent<MarkdownComponentProps> = createComponent();