+++ src/Blog/templates/Home/HomeTemplate.component.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { Head } from 'react-static';
 
 import { Link } from '../../../Shared/elements/Link/Link.component';
-import { Markdown } from '../../../Shared/elements/Markdown/Markdown.component';
-import { DefaultLayout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
+import { MarkdownBasic } from '../../../Shared/elements/MarkdownBasic/MarkdownBasic.component';
+import { DefaultLayout as Layout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
 import { ContentPage } from '../../../Shared/types/Page.model';
 
 import { Post } from '../../types/Post.model';
@@ -15,7 +15,7 @@ interface Props {
 
 export const HomeTemplate: React.StatelessComponent<Props> = props => {
   return (
-    <DefaultLayout>
+    <Layout>
       <Head title="Recent posts" meta={props.page.meta} />
       <h1 className="title-page" aria-label={props.page.title}>
         Latest Posts
@@ -24,10 +24,10 @@ export const HomeTemplate: React.StatelessComponent<Props> = props => {
         {props.posts.map((post: Post) => (
           <li key={post.path}>
             <Link to={post.path}>{post.title}</Link>
-            <Markdown text={post.abstract.source} stripped={true} />
+            <MarkdownBasic text={post.abstract.source} stripped={true} />
           </li>
         ))}
       </ul>
-    </DefaultLayout>
+    </Layout>
   );
 };