+++ src/Blog/templates/Post/PostTemplate.component.tsx
@@ -2,9 +2,10 @@ import * as React from 'react';
 import { Head } from 'react-static';
 
 import { Link } from '../../../Shared/elements/Link/Link.component';
-import { Markdown } from '../../../Shared/elements/Markdown/Markdown.component';
-import { DefaultLayout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
+import { DefaultLayout as Layout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
 
+import { PostAbstract } from '../../elements/PostAbstract/PostAbstract.component';
+import { PostContents } from '../../elements/PostContents/PostContents.component';
 import { Post } from '../../types/Post.model';
 
 import * as styles from './PostTemplate.module.scss';
@@ -15,12 +16,12 @@ interface Props {
 
 export const PostTemplate: React.StatelessComponent<Props> = props => {
   return (
-    <DefaultLayout>
+    <Layout>
       <Head title={props.post.title} meta={props.post.meta} />
       <Link to="/posts/">{'<'} Back</Link>
       <h1 className={styles.Post}>{props.post.title}</h1>
-      <Markdown text={props.post.abstract.source} />
-      <Markdown text={props.post.content.source} />
-    </DefaultLayout>
+      <PostAbstract text={props.post.abstract.source} />
+      <PostContents text={props.post.content.source} />
+    </Layout>
   );
 };