+++ src/Pages/templates/About/AboutTemplate.component.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Head } from 'react-static';
 
-import { Markdown } from '../../../Shared/elements/Markdown/Markdown.component';
-import { DefaultLayout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
+import { MarkdownBasic } from '../../../Shared/elements/MarkdownBasic/MarkdownBasic.component';
+import { DefaultLayout as Layout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
 import { ContentPage } from '../../../Shared/types/Page.model';
 
 interface Props {
@@ -11,10 +11,10 @@ interface Props {
 
 export const AboutTemplate: React.StatelessComponent<Props> = props => {
   return (
-    <DefaultLayout>
+    <Layout>
       <Head title={props.page.title} meta={props.page.meta} />
       <h1>{props.page.title}</h1>
-      <Markdown text={props.page.content.source} />
-    </DefaultLayout>
+      <MarkdownBasic text={props.page.content.source} />
+    </Layout>
   );
 };