+++ src/Shared/elements/MarkdownBasic/MarkdownBasic.component.tsx
@@ -1,0 +1,7 @@
+import * as markdown from '../../lib/markdown';
+
+import { basicComponentMap, MarkdownComponentProps, markdownFactory } from '../Markdown/Markdown.factory';
+
+const createComponent = () => markdownFactory(markdown.basic(), basicComponentMap);
+
+export const MarkdownBasic: React.StatelessComponent<MarkdownComponentProps> = createComponent();