+++ src/Site/templates/ErrorNotFound/NotFoundTemplate.component.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Head } from 'react-static';
 
-import { DefaultLayout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
-import { ContentPage } from '../../../Shared/types/Page.model';
+import { DefaultLayout as Layout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
+// import { ContentPage } from '../../../Shared/types/Page.model';
 
 import * as styles from './NotFoundTemplate.module.scss';
 
@@ -17,10 +17,10 @@ export const NotFoundTemplate: React.StatelessComponent<{}> = () => {
     }
   };
   return (
-    <DefaultLayout className={styles.layout}>
+    <Layout className={styles.layout}>
       <Head title={props.page.title} />
       <h1>{props.page.title}</h1>
       {/* <h1 className="title-page">Latest Posts {posts.length}</h1> */}
-    </DefaultLayout>
+    </Layout>
   );
 };