+++ static.config.js
@@ -1,7 +1,7 @@
-import configure from './.react-static/config.js';
-import watch from './.react-static/watch.js';
+import { configure } from './.react-static/config.js';
+import { watch } from './.react-static/watch.js';
 
-import routeBuilder from './src/routes';
+import { routeBuilder } from './src/routes';
 
 const stage = process.env.BUILD_STAGE || 'dev';